Selectors renamed to pages
